[FIX] timer logic

+++ src/routes/book/reading/useReadingSessionStore.ts
@@ -27,85 +27,85 @@ interface ReadingSessionState {
 }
 
 const useReadingSessionStore = create<ReadingSessionState>((set, get) => ({
-    timerInSeconds: parseInt(localStorage.getItem('timerSeconds') || '0', 10),
+    timerInSeconds: parseInt(localStorage.getItem('timerInSeconds') || '0', 10),
     isTimerOn: localStorage.getItem('isTimerOn') === 'true',
     isModalOpen: false,
     lastRecordedTime: localStorage.getItem('lastRecordedTime'),
+    bookId: localStorage.getItem('bookId') ? parseInt(localStorage.getItem('bookId')!, 10) : null,
+    readingSessionId: localStorage.getItem('readingSessionId') ? parseInt(localStorage.getItem('readingSessionId')!, 10) : null,
+    isEndModalOpen: false,
 
     startTimer: () => {
-        const now = new Date();
-        const lastRecordedTime = get().lastRecordedTime ? new Date(get().lastRecordedTime!) : new Date();
-        const differenceInSeconds = Math.floor((now.getTime() - lastRecordedTime.getTime()) / 1000);
+        const isAlreadyStarted = get().timerInSeconds > 0
+        if (isAlreadyStarted) return
+        const now = new Date()
 
-        set((state) => ({
-            timerInSeconds: state.timerInSeconds + differenceInSeconds,
-            isTimerOn: true,
-            lastRecordedTime: now.toISOString(),
-        }));
-
+        set({isTimerOn: true})
         localStorage.setItem('isTimerOn', 'true');
+
+        set({lastRecordedTime: now.toISOString()})
         localStorage.setItem('lastRecordedTime', now.toISOString());
+
+        set({timerInSeconds: 1})
+        localStorage.setItem('timerInSeconds', '1')
     },
 
     resumeTimer: () => {
         set({isTimerOn: true})
-
         localStorage.setItem('isTimerOn', 'true')
+
+        set({lastRecordedTime: new Date().toISOString()})
         localStorage.setItem('lastRecordedTime', new Date().toISOString());
     },
 
     pauseTimer: () => {
         set({isTimerOn: false})
-
         localStorage.setItem('isTimerOn', 'false')
-        localStorage.setItem('lastRecordedTime', new Date().toISOString());
+
+        set({lastRecordedTime: null})
+        localStorage.removeItem('lastRecordedTime')
     },
 
     resetTimer: () => {
         set({timerInSeconds: 0})
-        set({isTimerOn: false})
-        set({bookId: null})
-        set({readingSessionId: null})
-        set({lastRecordedTime: null})
-        set({isModalOpen: false})
-        set({isEndModalOpen: false})
-
         localStorage.removeItem('timerSeconds')
+
+        set({isTimerOn: false})
         localStorage.removeItem('isTimerOn')
-        localStorage.removeItem('lastRecordedTime')
+
+        set({bookId: null})
         localStorage.removeItem('bookId')
+
+        set({readingSessionId: null})
         localStorage.removeItem('readingSessionId')
+
+        set({lastRecordedTime: null})
+        localStorage.removeItem('lastRecordedTime')
+
+        set({isModalOpen: false})
+
+        set({isEndModalOpen: false})
     },
 
     incrementTimer: () => {
         if (!get().isTimerOn) return
 
-        const now = new Date();
-        const lastRecordedTime = get().lastRecordedTime ? new Date(get().lastRecordedTime!) : new Date();
-        const differenceInSeconds = Math.floor((now.getTime() - lastRecordedTime.getTime()) / 1000);
+        const now = new Date()
+        const lastRecordedTime = get().lastRecordedTime ? new Date(get().lastRecordedTime!) : now
+        const differenceInSeconds = Math.floor((now.getTime() - lastRecordedTime.getTime()) / 1000)
+        const differenceToAdd = differenceInSeconds < 3 ? 0 : differenceInSeconds
 
-        set((state) => ({
-            timerInSeconds: state.timerInSeconds + differenceInSeconds,
-            lastRecordedTime: now.toISOString(),
-        }));
+        set({lastRecordedTime: now.toISOString()})
+        localStorage.setItem('lastRecordedTime', now.toISOString())
+
+        set({timerInSeconds: get().timerInSeconds + differenceToAdd + 1})
+        localStorage.setItem('timerInSeconds', (get().timerInSeconds + differenceToAdd + 1).toString())
     },
 
     isShowingTimer: () => {
         return get().bookId != null && get().readingSessionId != null
     },
 
-    openModal: (bookId: number) => {
-        set({isModalOpen: true})
-        set({bookId: bookId})
-        localStorage.setItem('bookId', bookId.toString())
-    },
-
-    closeModal: () => {
-        set({isModalOpen: false})
-    },
-
-    bookId: localStorage.getItem('bookId') ? parseInt(localStorage.getItem('bookId')!, 10) : null,
-    readingSessionId: localStorage.getItem('readingSessionId') ? parseInt(localStorage.getItem('readingSessionId')!, 10) : null,
     setBookId: (bookId: number) => {
         localStorage.setItem('bookId', bookId.toString())
         set({bookId})
@@ -115,7 +115,16 @@ const useReadingSessionStore = create<ReadingSessionState>((set, get) => ({
         set({readingSessionId})
     },
 
-    isEndModalOpen: false,
+    openModal: (bookId: number) => {
+        set({isModalOpen: true})
+
+        set({bookId: bookId})
+        localStorage.setItem('bookId', bookId.toString())
+    },
+    closeModal: () => {
+        set({isModalOpen: false})
+    },
+
     openEndModal: () => {
         set({isEndModalOpen: true})
     },
